+++ backend/src/validators/glossary.js
@@ -1,0 +1,9 @@
+const { getResponse } = require("./common_schema.js");
+
+const getAll = {
+  response: getResponse(),
+};
+
+module.exports = {
+  getAll,
+};